ENH: Resolve circular dependancies by using separate module.
